+++ src/main/java/com/mbc/sharetravel_spring/repository/PostRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mbc.sharetravel_spring.posts.TravelBoard;
 
-public interface PostRepository extends JpaRepository<TravelBoard, Long> {
+public interface PostRepository extends JpaRepository<TravelBoard, Integer> {
 	List<TravelBoard> findAll();
 }
